Rename config.rs to a more accurate name
